backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
pydantic[email-validator]==2.9.2
email-validator==2.1.0
pydantic-settings==2.5.2
python-socketio==5.11.4
apscheduler==3.10.4
alembic==1.13.3
greenlet==3.1.1
